+++ src/hooks/useDragControl.ts
@@ -1,5 +1,5 @@
 import { useEffect, useContext } from "react";
-import { DragControlContext } from "../Tree";
+import { DragControlContext } from "../providers";
 
 /**
  * This is a hook to allow text selection by mouse in the text input area in a node.
@@ -9,19 +9,23 @@ export const useDragControl = (ref: React.RefObject<HTMLElement>): void => {
   const dragControlContext = useContext(DragControlContext);
 
   const lock = (e: Event) => {
-    const target = e.target as Element;
-    const tagName = target.tagName.toLowerCase();
+    const { target } = e;
 
-    if (tagName === "input" || tagName === "textarea") {
+    if (
+      target instanceof HTMLInputElement ||
+      target instanceof HTMLTextAreaElement
+    ) {
       dragControlContext.lock();
     }
   };
 
   const unlock = (e: Event) => {
-    const target = e.target as Element;
-    const tagName = target.tagName.toLowerCase();
+    const { target } = e;
 
-    if (tagName === "input" || tagName === "textarea") {
+    if (
+      target instanceof HTMLInputElement ||
+      target instanceof HTMLTextAreaElement
+    ) {
       dragControlContext.unlock();
     }
   };