+++ src/providers/index.tsx
@@ -1,0 +1,25 @@
+import React, { PropsWithChildren, ReactElement } from "react";
+import { DndProvider } from "react-dnd-multi-backend";
+import HTML5toTouch from "react-dnd-multi-backend/dist/cjs/HTML5toTouch";
+import { TreeProvider, TreeContext } from "./TreeProvider";
+import { DragControlProvider, DragControlContext } from "./DragControlProvider";
+import { PlaceholderProvider, PlaceholderContext } from "./PlaceholderProvider";
+import { TreeProps, OpenIdsHandlers } from "../types";
+
+type Props<T> = PropsWithChildren<
+  TreeProps<T> & {
+    treeRef: React.ForwardedRef<OpenIdsHandlers>;
+  }
+>;
+
+export const Providers = <T extends unknown>(props: Props<T>): ReactElement => (
+  <TreeProvider {...props}>
+    <DragControlProvider>
+      <PlaceholderProvider>
+        <DndProvider options={HTML5toTouch}>{props.children}</DndProvider>
+      </PlaceholderProvider>
+    </DragControlProvider>
+  </TreeProvider>
+);
+
+export { TreeContext, DragControlContext, PlaceholderContext };