+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export { getTreeItem } from "./getTreeItem";
 export { isAncestor } from "./isAncestor";
 export { isDroppable } from "./isDroppable";
 export { mutateTree } from "./mutateTree";
+export { getDropTarget } from "./getDropTarget";